fix: can create HB signal if selected any necessary field
