+++ src/ducks/HistoricalBalance/balanceView/BalanceChartHeader.js
@@ -4,7 +4,7 @@ import {
   ETH_WALLET_AMOUNT_UP,
   ETH_WALLET_METRIC
 } from '../../Signals/utils/constants'
-import { mapToOptions } from '../../Signals/utils/utils'
+import { mapToOptions } from '../../../utils/select/utils'
 import styles from './BalanceView.module.scss'
 
 const BalanceChartHeader = ({ address, assets, children }) => {
@@ -13,7 +13,7 @@ const BalanceChartHeader = ({ address, assets, children }) => {
       <div className={styles.addTrigger}>
         <SignalMasterModalForm
           label='Generate signal'
-          enabled={address && assets && assets.length > 0}
+          enabled={address || (assets && assets.length > 0)}
           canRedirect={false}
           metaFormSettings={{
             target: {
@@ -25,7 +25,7 @@ const BalanceChartHeader = ({ address, assets, children }) => {
             type: {
               value: { ...ETH_WALLET_AMOUNT_UP }
             },
-            ethAddress: address
+            ethAddress: mapToOptions(address)
           }}
           buttonParams={{
             variant: 'ghost',
