+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/TriggerFormHistoricalBalance.js
@@ -134,12 +134,7 @@ const TriggerFormHistoricalBalance = ({
       allErc20Projects &&
         allErc20Projects.length &&
         !erc20List.length &&
-        setErc20([
-          ...allErc20Projects,
-          {
-            slug: ETHEREUM
-          }
-        ])
+        setErc20(allErc20Projects)
       allProjects && allProjects.length && setAll(allProjects)
       assets && assets.length && setHeldAssets(assets)
     },
@@ -220,6 +215,13 @@ const mapDataToProps = ({
   ownProps
 }) => {
   const { data = {} } = ownProps
+
+  if (allErc20Projects && allErc20Projects.length > 0) {
+    allErc20Projects.push({
+      slug: ETHEREUM
+    })
+  }
+
   return {
     ...ownProps,
     data: {
