+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/projectsSelector/ProjectsList.module.scss
@@ -34,7 +34,7 @@
 
 .balance {
   color: var(--jungle-green);
-  margin-left: 5px;
+  margin-right: 5px;
 }
 
 .checkbox {
